+++ src/main/java/aibe/hosik/profile/repository/ProfileRepository.java
@@ -1,5 +1,6 @@
-package aibe.hosik.profile;
+package aibe.hosik.profile.repository;
 
+import aibe.hosik.profile.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
