+++ src/main/java/aibe/hosik/review/controller/ReviewController.java
@@ -1,7 +1,8 @@
-package aibe.hosik.review;
+package aibe.hosik.review.controller;
 
 import aibe.hosik.review.dto.ReviewRequest;
 import aibe.hosik.review.dto.ReviewResponse;
+import aibe.hosik.review.service.ReviewService;
 import aibe.hosik.user.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
