+++ src/main/java/aibe/hosik/review/dto/ReviewResponse.java
@@ -1,9 +1,8 @@
 package aibe.hosik.review.dto;
 
-import aibe.hosik.post.dto.PostDetailDTO;
 import aibe.hosik.post.entity.Post;
 import aibe.hosik.profile.dto.ProfileResponse;
-import aibe.hosik.review.Review;
+import aibe.hosik.review.entity.Review;
 
 public record ReviewResponse(
     Long id,
