+++ src/main/java/aibe/hosik/review/entity/Review.java
@@ -1,4 +1,4 @@
-package aibe.hosik.review;
+package aibe.hosik.review.entity;
 
 import aibe.hosik.common.TimeEntity;
 import aibe.hosik.post.entity.Post;
