+++ src/main/java/aibe/hosik/review/repository/ReviewRepository.java
@@ -1,9 +1,9 @@
-package aibe.hosik.review;
+package aibe.hosik.review.repository;
 
+import aibe.hosik.review.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
   List<Review> findAllByRevieweeId(Long userId);
