+++ src/main/java/aibe/hosik/user/UserService.java
@@ -3,9 +3,8 @@ package aibe.hosik.user;
 import aibe.hosik.auth.dto.SignUpRequest;
 import aibe.hosik.auth.dto.PasswordChangeRequest;
 import aibe.hosik.auth.dto.SocialLoginRequest;
-import aibe.hosik.profile.Profile;
-import aibe.hosik.profile.ProfileRepository;
-import aibe.hosik.profile.ProfileService;
+import aibe.hosik.profile.entity.Profile;
+import aibe.hosik.profile.repository.ProfileRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
